+++ src/components/manage/UniversalLink/UniversalLink.jsx
@@ -9,6 +9,7 @@ import { Link } from 'react-router-dom';
 import { useSelector } from 'react-redux';
 import { flattenToAppURL, isInternalURL } from '@plone/volto/helpers/Url/Url';
 import URLUtils from '@plone/volto/components/manage/AnchorPlugin/utils/URLUtils';
+import config from '@plone/volto/registry';
 
 const UniversalLink = ({
   href,
@@ -30,7 +31,10 @@ const UniversalLink = ({
     }
   }
 
-  const isExternal = !isInternalURL(url);
+  const isBlacklisted = (config.settings.internalUrlBlacklist ?? []).includes(
+    flattenToAppURL(url),
+  );
+  const isExternal = !isInternalURL(url) || isBlacklisted;
   const isDownload = (!isExternal && url.includes('@@download')) || download;
 
   return isExternal ? (